controllers: move unit test funcs
